# WIP: gen/FromWire: Allocate pointer fields once