--- a/Makefile
+++ b/Makefile
@@ -134,3 +134,4 @@ cover:
 .PHONY: clean
 clean:
 	go clean
+	rm $(GOBIN)/thriftrw