--- a/gen/internal/tests/non_hyphenated/non_hyphenated.go
+++ b/gen/internal/tests/non_hyphenated/non_hyphenated.go
@@ -56,6 +56,9 @@ func (v *First) ToWire() (wire.Value, error) {
 //   }
 //   return &v, nil
 func (v *First) FromWire(w wire.Value) error {
+	var ptrFields struct {
+	}
+	_ = ptrFields
 
 	for _, field := range w.GetStruct().Fields {
 		switch field.ID {
@@ -146,6 +149,9 @@ func (v *Second) ToWire() (wire.Value, error) {
 //   }
 //   return &v, nil
 func (v *Second) FromWire(w wire.Value) error {
+	var ptrFields struct {
+	}
+	_ = ptrFields
 
 	for _, field := range w.GetStruct().Fields {
 		switch field.ID {