--- a/gen/internal/tests/set_to_slice/set_to_slice.go
+++ b/gen/internal/tests/set_to_slice/set_to_slice.go
@@ -432,6 +432,10 @@ func _StringListList_Read(w wire.Value) (StringListList, error) {
 //   }
 //   return &v, nil
 func (v *Bar) FromWire(w wire.Value) error {
+	var ptrFields struct {
+	}
+	_ = ptrFields
+
 	var err error
 
 	requiredInt32ListFieldIsSet := false
@@ -985,6 +989,10 @@ func (v *Foo) ToWire() (wire.Value, error) {
 //   }
 //   return &v, nil
 func (v *Foo) FromWire(w wire.Value) error {
+	var ptrFields struct {
+	}
+	_ = ptrFields
+
 	var err error
 
 	stringFieldIsSet := false