… gen/internal/tests/services/services.go
… gen/internal/tests/structs/structs.go